Add message when more than 24 results are found
When over 24 results are found on the Cycle Babac website, only the first 24
are tabulated. However the user will now be provided with a link to get to
the full list.

Bug correction: if a product does not have an associated SKU, it now returns
N/A for the SKU.