Add JSON endpoint to search for a single SKU
This endpoint will allow applications to request information about a
single product by specifying its SKU.  Applications must issue a GET
request on /json/<sku>, where sku must be of the format XX-XXX, where X
is a digit.

If no product is found, the 404 status code is returned.  Otherwise, 200
is returned, with a json structure containing the product details.